data_pipeline/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
apprise==1.9.2
asgi-lifespan==2.1.0
asttokens==3.0.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
coolname==2.2.0
cryptography==44.0.2
dateparser==1.2.1
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
docker==7.1.0
exceptiongroup==1.2.2
exchange_calendars==4.10
executing==2.2.0
fastapi==0.115.12
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.3.0
graphviz==0.20.3
greenlet==3.1.1
griffe==1.7.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.12.2
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kafka-python==2.1.4
korean-lunar-calendar==0.3.1
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.2.0
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==2.2.4
oauthlib==3.2.2
opentelemetry-api==1.31.1
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pandas_market_calendars==4.6.1
parso==0.8.4
pathspec==0.12.1
peewee==3.17.9
pendulum==3.0.0
pexpect==4.9.0
pi==0.1.2
platformdirs==4.3.7
polygon-api-client==1.14.4
prefect==3.2.15
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.1
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.33.0
Pygments==2.19.1
pyluach==2.2.0
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
python-socks==2.7.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.1
ta==0.11.0
text-unidecode==1.3
time-machine==2.16.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
traitlets==5.14.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.3.0
uv==0.6.10
uvicorn==0.34.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.2
wrapt==1.17.2
yarl==1.18.3
yfinance==0.2.55
zipp==3.21.0
